requirements.txt
requests
azure-mgmt-billing
azure-mgmt-consumption
azure-identity
azure-mgmt-costmanagement